Working discriminator for AI vs Human text
Getting 99.96% accuracy on the first set of essay data after 1 epoch.

Manual testing with random paragraphs from essays and books found on my
computer seem to be producing very good results, with few exceptions.

Exceptions seem to be false positives, not false negatives (i.e. it
thinks the human writing is AI, not the other way around), even when I
ask the AI to write like an 8th grader.

Perhaps there is a ROC that we can tune to avoid false positives.