DeBERTa LoRA: Temperature scaling
In order to reduce the model's overconfidence, I added a temperature
quotient to the logits.

See On Calibration of Modern Neural Networks
(https://arxiv.org/abs/1706.04599) and the example implementation:
https://github.com/gpleiss/temperature_scaling

In the future I may add a load/save feature for this, but for now I will
simply copy-paste the value into my inference scripts.